dom/workers/test/test_chromeWorker.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Test for DOM Worker Threads</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

  try {
    var worker = new ChromeWorker("simpleThread_worker.js");
    ok(false, "ChromeWorker constructor should be blocked!");
  }
  catch (e) {
    ok(true, "ChromeWorker constructor wasn't blocked!");
  }

</script>
</pre>
</body>
</html>
